fix(lib>rtnt>core>packet): Fixed invalid cast size.

+++ lib/rtnt/include/rtnt/Core/Packet.hpp
@@ -40,7 +40,7 @@ namespace packet
         uint32_t acknowledgeId          = 0; ///< Sequence ID of the latest packet received
         uint32_t acknowledgeBitfield    = 0; ///< Bitmask of the previous 32 received packets relative to acknowledge ID
         uint16_t messageId              = 0x0; ///< Command type (user-defined)
-        uint8_t  flags                  = static_cast<uint16_t>(Flag::kUnreliable); ///< Reliability flags (cf. packet::Flag)
+        uint8_t  flags                  = static_cast<uint8_t>(Flag::kUnreliable); ///< Reliability flags (cf. packet::Flag)
         uint16_t packetSize             = 0; ///< Size of the payload
         uint32_t checksum               = 0; ///< CRC32 checksum to avoid corruption
